requirements.txt
allure-pytest
pydantic-settings
pytest
requests
urllib3
allure-python-commons